Fix alpha beldning and cull back faces

+++ src/scripts/shaders/defaultFrag.ts
@@ -10,7 +10,7 @@ uniform vec2 u_resolution;
 uniform vec4 u_mouse;
 uniform float u_time;
 uniform float u_timeDelta;
-uniform float u_frameNumber;
+uniform int u_frameNumber;
 
 void main() {
     fragColor = vec4((FragPosition + 1.0) / 2.0, 1.0);
